+++ client/components/widgets/SelectUser.vue
@@ -10,13 +10,30 @@
       <v-card>
         <v-card-title class="headline grey lighten-2">{{ $t('Select user')}}</v-card-title>
         <v-card-text>
-          <v-tabs>
-            <v-tab ripple>{{ $t('Available users') }}</v-tab>
-            <v-tab ripple v-if="isAdmin">{{ $t('Find')}}</v-tab>
-            <v-tab-item>
+          <v-tabs ref="tabs" v-if="active">
+            <v-tab v-if="!hideProject && project">{{ $t('Project') }}</v-tab>
+            <v-tab v-if="project && project.organizationId">{{ $t('Organization') }}</v-tab>
+            <v-tab v-if="isAdmin">{{ $t('Find')}}</v-tab>
+            <v-tab-item v-if="!hideProject && project">
               <div class="flex-container">
                 <v-list class="flex1" dense subheader>
-                  <template v-for="user in availableUsers">
+                  <template v-for="user in projectUsers">
+                    <v-list-tile :key="user._id" avatar @click="selectUser(user)">
+                      <v-list-tile-avatar :color="isOnline(user)">
+                        <span class>{{ formatUserLetters(user) }}</span>
+                      </v-list-tile-avatar>
+                      <v-list-tile-content class="pointer">
+                        <v-list-tile-title>{{ formatUser(user) }}</v-list-tile-title>
+                      </v-list-tile-content>
+                    </v-list-tile>
+                  </template>
+                </v-list>
+              </div>
+            </v-tab-item>
+            <v-tab-item v-if="project && project.organizationId">
+              <div class="flex-container">
+                <v-list class="flex1" dense subheader>
+                  <template v-for="user in organizationUsers">
                     <v-list-tile :key="user._id" avatar @click="selectUser(user)">
                       <v-list-tile-avatar :color="isOnline(user)">
                         <span class>{{ formatUserLetters(user) }}</span>
@@ -39,6 +56,7 @@
                     v-model="search"
                     append-icon="search"
                     clearable
+                    autofocus
                     v-on:input="debouncedFilter"
                   ></v-text-field>
                 </div>
@@ -142,6 +160,10 @@ export default {
   },
   props: {
     active: Boolean,
+    hideProject: { 
+      type :Boolean,
+      default: false
+    },
     project: Object,
     isAdmin: {
       type: Boolean,
@@ -150,6 +172,7 @@ export default {
   },
   data() {
     return {
+      selectedTab: 0,
       search: "",
       debouncedFilter: null,
       users: [],
@@ -163,8 +186,13 @@ export default {
       }
     };
   },
-  meteor: {
-    availableUsers() {
+  computed: {
+     projectUsers() {
+      if (this.project) {
+        return Meteor.users.find({ _id: { $in: this.project.members } });
+      }
+    },
+    organizationUsers() {
       if (this.project && this.project.organizationId) {
         const organization = Organizations.findOne(this.project.organizationId);
         if (organization) {
@@ -172,7 +200,6 @@ export default {
           return Meteor.users.find({ _id: { $in: members } });
         }
       }
-      return Meteor.users.find();
     }
   },
   methods: {