+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Display previous members in project filter toolbar
+- Select user widget now display 3 tabs: project members/organization members/find
 
 ### Removed